refactor: 구조 정리-커뮤니티
